Refactored and optimized the code :boom: :rocket:
