+++ package.json
@@ -15,6 +15,7 @@
   "license": "ISC",
   "dependencies": {
     "browserify": "^13.0.0",
+    "iterate-object": "^1.3.1",
     "watchify": "^3.7.0"
   },
   "repository": {
