+++ staticAssets/index.html
@@ -12,13 +12,13 @@
         <div id='content'>
             <h1>volodymyr zabeida</h1>
              <ul id='menu'>
-                <li><a id='drawings'></a></li>
-                <li><a id='etchings'></a></li>
-                <li><a id='monotypes'></a></li>
-                <li><a id='paintings'></a></li>
-                <li><a id='photography'></a></li>
-                <li><a id='models'></a></li>
-                <li><a id='biography'></a></li>
+                <li><a class="selected" data-category='drawings'></a></li>
+                <li><a data-category='etchings'></a></li>
+                <li><a data-category='monotypes'></a></li>
+                <li><a data-category='paintings'></a></li>
+                <li><a data-category='photography'></a></li>
+                <li><a data-category='models'></a></li>
+                <li><a data-category='biography'></a></li>
             </ul>
 
             <div id='display'>
@@ -33,19 +33,3 @@
         <script src='/scripts/bundle.js'></script>
     </body>
 </html>
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
