+++ staticAssets/scripts/ui/index.js
@@ -1,0 +1,11 @@
+module.exports = {
+    menu: document.getElementById('menu'),
+    links: document.querySelectorAll('a[data-category]'),
+    langButton: document.querySelector('#language a'),
+    container: document.getElementById('image'),
+    contentHeight: 500,
+    content: document.getElementsByClassName('content'),
+    descriptionContainer: document.getElementById('description'),
+    titleContainer: document.getElementById('title'),
+    contentDivs: []
+};
